ProSimNez_simulator - コピー/backend/requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# シミュレーション
simpy==4.0.2

# データ処理
pandas==2.1.3
numpy==1.26.2

# データベース
sqlalchemy==2.0.23
alembic==1.12.1

# 非同期処理
celery==5.3.4
redis==5.0.1

# セキュリティ
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# テスト
pytest==7.4.3
pytest-asyncio==0.21.1

# その他
python-dotenv==1.0.0
pyyaml==6.0.1
setuptools==69.0.2